Fixing over-encoding of paths for iOS device debugging
When attempting to determine which webview to attach to we URL encode the package name to deal with spaces and other special characters. However prior to this fix we pass the whole on-device path for device debugging, and we would URL encode the path separators. Now we only match on, and encode, the app folder itself.